% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_apis.R
\docType{data}
\name{API_KEY_FREDR}
\alias{API_KEY_FREDR}
\title{API key for FRED database from St. Louis Fed}
\format{
An object of class \code{character} of length 1.
}
\usage{
API_KEY_FREDR
}
\description{
API key for FRED database from St. Louis Fed
}
\keyword{datasets}
