% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomberg.R
\name{bdh}
\alias{bdh}
\title{Bloomberg history download}
\usage{
bdh(
  tickers,
  fields = "PX_LAST",
  start_date = "-7CD",
  end_date = NULL,
  period = "DAILY",
  conn = NULL
)
}
\description{
Bloomberg history download
}
