% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_income.R
\name{bond_price}
\alias{bond_price}
\title{Calculates bond price}
\usage{
bond_price(yield, coupon, maturity, coupfreq = 1, redemption = 100)
}
\arguments{
\item{yield}{annual yield to maturity, expressed as .05 = 5\%}

\item{coupon}{annual coupon, expressed as .05 = 5\%}

\item{maturity}{maturity in years}

\item{coupfreq}{number of coupons per year, default = 1}

\item{redemption}{redemption price, default = 100}
}
\description{
Calculates bond price
}
