% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_functions.R
\name{chart_plot}
\alias{chart_plot}
\title{Main chart plot function}
\usage{
chart_plot(
  xdata = c(0, 1),
  ydata = c(0, 1),
  revx = F,
  bgcolor = caim::caim_colors("lightash"),
  textcolor = caim::caim_colors("dark1"),
  showxvalues = T,
  showyvalues = T,
  showxaxis = F,
  showyaxis = F,
  showgrid = T,
  xat = NULL,
  yat = NULL,
  xfmt = c("dec", 2),
  yfmt = c("dec", 2),
  xaxistext = NULL,
  yaxistext = NULL,
  source = NULL,
  xlab = NA,
  ylab = NA,
  ...
)
}
\description{
Main chart plot function
}
