% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{cor2cov}
\alias{cor2cov}
\title{Calculate covariance matrix from correlation matrix and vector of standard deviations}
\usage{
cor2cov(cormtx, stdevs)
}
\description{
Calculate covariance matrix from correlation matrix and vector of standard deviations
}
