% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{date_range}
\alias{date_range}
\title{Finds date range of data series in long data.table}
\usage{
date_range(data, keycols, datecol = "date")
}
\description{
Finds date range of data series in long data.table
}
