% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_functions.R
\name{heatmap_color}
\alias{heatmap_color}
\title{Heatmap color}
\usage{
heatmap_color(
  x,
  range = c(-1, 1),
  color_lo = caim::caim_colors("darkblue"),
  color_mid = caim::caim_colors("white"),
  color_hi = caim::caim_colors("darkred")
)
}
\description{
Heatmap color
}
