% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_ice_historical_data}
\alias{import_ice_historical_data}
\title{Imports data from ICE historical data download .csv files}
\usage{
import_ice_historical_data(filenames, field_key = NULL)
}
\arguments{
\item{filenames}{character vector}

\item{field_key}{if NULL, function will use default values}
}
\description{
From Bloomberg IND13
}
