% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{INF}
\alias{INF}
\title{Maximum double value}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
INF
}
\description{
Replacement for Inf when functions can't handle Inf (like quadprog, apparently)
}
\keyword{datasets}
