% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{is_ymd}
\alias{is_ymd}
\title{Checks to see whether a character string can be formatted as YYYY-MM-DD}
\usage{
is_ymd(date)
}
\description{
Checks to see whether a character string can be formatted as YYYY-MM-DD
}
