% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_validation.R
\name{k_fold_ix}
\alias{k_fold_ix}
\title{Generates k-fold selection index}
\usage{
k_fold_ix(n, k = 10, seed = NA)
}
\arguments{
\item{n}{number of observations.}

\item{k}{number of folds. default = 10}

\item{seed}{seed for random number generator. default = NA}
}
\description{
Generates k-fold selection index
}
