% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm_quad}
\alias{lm_quad}
\title{Constrained weighted linear regression via quadratic programming}
\usage{
lm_quad(
  y,
  X,
  Aeq = NULL,
  beq = NULL,
  A = NULL,
  b = NULL,
  lb = NULL,
  ub = NULL,
  wts = NULL,
  fixed_coefs = NULL
)
}
\arguments{
\item{y}{vector of Y values}

\item{X}{feature matrix}

\item{Aeq}{Amat for equality constraints. numconstraints x numX}

\item{beq}{bvec for inequality constraints}

\item{A}{Amat for >= inequality constraints}

\item{b}{bvec for >= inequality constraints}

\item{lb}{vector of lower bounds}

\item{ub}{vector of upper bounds}

\item{wts}{observation weightings. default=equal-weighted}

\item{fixed_coefs}{if !is.null, does not run regression, but builds model with given coefs}
}
\description{
Linear regression with weights and constraints can be expressed as a
quadratic programming problem. See, for example,
https://en.wikipedia.org/wiki/Weighted_least_squares
}
