% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_income.R
\name{macaulay_duration}
\alias{macaulay_duration}
\title{Calculates Macaulay duration}
\usage{
macaulay_duration(yield, coupon, maturity, coupfreq)
}
\arguments{
\item{yield}{annual yield to maturity, expressed as .05 = 5\%}

\item{coupon}{annual coupon, expressed as .05 = 5\%}

\item{maturity}{maturity in years}

\item{coupfreq}{number of coupons per year, default = 1}
}
\description{
Calculates Macaulay duration
}
