% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{max_dates}
\alias{max_dates}
\title{Finds latest dates of data series in long data.table}
\usage{
max_dates(data, keycols, datecol = "date")
}
\description{
Finds latest dates of data series in long data.table
}
