% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{min_dates}
\alias{min_dates}
\title{Finds earliest dates of data series in long data.table}
\usage{
min_dates(data, keycols, datecol = "date")
}
\description{
Finds earliest dates of data series in long data.table
}
