% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{month_end_dates}
\alias{month_end_dates}
\title{Returns last date in each month for date series}
\usage{
month_end_dates(dates)
}
\arguments{
\item{dates}{vector of dates in "YYYY-MM-DD" format}
}
\description{
Returns last date in each month for date series
}
