% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{monthly_data}
\alias{monthly_data}
\title{Extracts month end series from long data.table}
\usage{
monthly_data(data, keycols, datecol = "date")
}
\description{
Extracts month end series from long data.table
}
