% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{NEG_INF}
\alias{NEG_INF}
\title{Minimum double value}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
NEG_INF
}
\description{
Replacement for -Inf when functions can't handle Inf (like quadprog, apparently)
}
\keyword{datasets}
