% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson_siegel.R
\name{ns_coefs2yields}
\alias{ns_coefs2yields}
\title{Generates yield curve from Nelson-Siegel coefficients}
\usage{
ns_coefs2yields(
  coefs = NULL,
  mats = c(0.0833, seq(0.25, 30, 0.25)),
  beta0 = NA,
  beta1 = NA,
  beta2 = NA,
  lambda = NA
)
}
\description{
Generates yield curve from Nelson-Siegel coefficients
}
