% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson_siegel.R
\name{ns_fwd}
\alias{ns_fwd}
\title{Calculates Nelson-Siegel lambda that maximizes beta2 loading at mat}
\usage{
ns_fwd(coefs, timefwd)
}
\description{
Calculates Nelson-Siegel lambda that maximizes beta2 loading at mat
}
