% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson_siegel.R
\name{ns_lambda2mat}
\alias{ns_lambda2mat}
\title{Calculates Nelson-Siegel lambda that maximizes beta2 loading at mat}
\usage{
ns_lambda2mat(lambda, matrange = c(0, 30))
}
\description{
Calculates Nelson-Siegel lambda that maximizes beta2 loading at mat
}
