% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson_siegel.R
\name{ns_mat2lambda}
\alias{ns_mat2lambda}
\title{Calculates maturity where Nelson-Siegel beta2 loading is maximized for given lambda}
\usage{
ns_mat2lambda(mat, lambdarange = c(0, 1000))
}
\description{
Calculates maturity where Nelson-Siegel beta2 loading is maximized for given lambda
}
