% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelson_siegel.R
\name{ns_yields2coefs}
\alias{ns_yields2coefs}
\title{Generates Nelson-Siegel coefficients from maturities and yields}
\usage{
ns_yields2coefs(
  mats,
  yields,
  lambda = 0.7173239,
  wts = rep(1/length(mats), length(mats))
)
}
\description{
Generates Nelson-Siegel coefficients from maturities and yields
}
