% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_functions.R
\name{number_format}
\alias{number_format}
\title{format number to text}
\usage{
number_format(num, fmt = c("pct", 2))
}
\arguments{
\item{num}{number or vector of numbers to format}

\item{fmt}{format to apply. Default = c("pct", 2). Acceptable formats = "pct", "dec", "dk"
(divided by 1000)}
}
\description{
format number to text
}
\examples{
number_format(.0023, c("pct", 2))
number_format(c(.016784, .002), c("dec", 4))
number_format(1234567.89, c("dk", 0))
}
