% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{opt_cvar}
\alias{opt_cvar}
\title{Conditional Value at Risk portfolio optimization}
\usage{
opt_cvar(
  assetrets,
  exprets = apply(assetrets, 2, mean),
  confidence = 0.95,
  opttype = "tgtret",
  target = 0,
  name = "Portfolio",
  lb = rep(0, length(exprets)),
  ub = rep(1, length(exprets)),
  bounds = NA,
  portwt = 1
)
}
\arguments{
\item{assetrets}{numobs x numassets numeric matrix of asset returns}

\item{exprets}{vector of expected returns. Default - calculated automatically from
\code{assetrets}}

\item{confidence}{numeric CVaR confidence level. Default = .95}

\item{opttype}{character optimization type. Default = "tgtret". Valid options include
"minrisk", "tgtret", "tgtrisk", "maxret"}

\item{target}{numeric target value, depending on optimization type. Default = 0}

\item{name}{character name of optimized portfolio}
}
\value{
Optimization results list (one portfolio) including $name, $solved, $wts, $exprets, $ret
 $vol, $var, $var.model, $cvar, $cvar.model, $summary, $model, objfn
}
\description{
Portfolio optimization using CVaR risk constraints. Linear programming a la Uryasev
}
\examples{
opt_cvar(oi::simrets, opttype="tgtrisk", target=0)
opt_cvar(oi::simrets, opttype="tgtret", target=.03, bounds=oi::samplebounds$bounds[[2]])
}
\seealso{
\code{\link{opt_mv}}: mean-variance portfolio optimization \cr
\code{\link{frontier_cvar}}: CVaR efficient frontier \cr
}
