% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{quarter_end_dates}
\alias{quarter_end_dates}
\title{Returns last date in each quarter for date series}
\usage{
quarter_end_dates(dates)
}
\description{
Returns last date in each quarter for date series
}
