% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_ix}
\alias{sim_ix}
\title{Bootstrap simulation "selection index"}
\usage{
sim_ix(num_obs, num_sims, block_size = 1, prob = NULL, seed = NULL)
}
\description{
Bootstrap simulation "selection index"
}
