% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_period_ix}
\alias{sim_period_ix}
\title{Bootstrap simulation "selection table"}
\usage{
sim_period_ix(
  num_sims,
  num_periods,
  num_obs,
  block_size = 1,
  prob = NULL,
  seed = NULL
)
}
\description{
Bootstrap simulation "selection table"
}
