% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{wtd_cor}
\alias{wtd_cor}
\title{Calculate weighted correlation}
\usage{
wtd_cor(x, y = NULL, wts = NULL, population = F)
}
\arguments{
\item{x}{vector or matrix of data}

\item{y}{vector of data. default = NULL.}

\item{wts}{vector of weights. NULL will produce equally-weighted results. wts will be normalized
to sum to 1. default = NULL}

\item{population}{F will return sample correlation, T will return population correlation default=F}
}
\value{
correlation matrix if y == NULL, otherwise correlation between x and y.
}
\description{
Calculate weighted correlation
}
