% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{wtd_mean}
\alias{wtd_mean}
\title{Calculate weighted mean}
\usage{
wtd_mean(x, wts = NULL)
}
\arguments{
\item{x}{vector or matrix of data}

\item{wts}{vector of wts. NULL will produce equally-weighted results. wts will be normalized
to sum to 1. default = NULL}
}
\description{
Calculate weighted mean
}
