% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{wtd_sd}
\alias{wtd_sd}
\title{Calculate weighted standard deviation}
\usage{
wtd_sd(x, wts = NULL, population = F)
}
\arguments{
\item{x}{vector or matrix of data}

\item{wts}{vector of weights. NULL will produce equally-weighted results. wts will be normalized
to sum to 1. default = NULL}

\item{population}{F will return sample std dev, T will return population std dev. default=F}
}
\description{
Calculate weighted standard deviation
}
