% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{wtd_variance}
\alias{wtd_variance}
\title{Calculate weighted variance}
\usage{
wtd_variance(x, wts = NULL, population = F)
}
\arguments{
\item{x}{vector or matrix of data}

\item{wts}{vector of weights. NULL will produce equally-weighted results. wts will be normalized
to sum to 1. default = NULL}

\item{population}{F will return sample variance, T will return population variance. default=F}
}
\description{
Calculate weighted variance
}
